+++ HISTORY.rst
@@ -12,6 +12,8 @@ Unreleased
 **Bugfixes**
 * Fix checksum generation for colis pharmacy
 
+* On export of stock exported life_date is now based on lot with quantities > 0
+
 **Build**
 
 **Documentation**